# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GophKeeper HTTP API
  version: 1.0.0
  description: >
    Public HTTP API for authentication and secrets sync. All secret payloads
    are encrypted client-side; the server stores and returns opaque payloads.
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
paths:
  /health:
    get:
      summary: Health check
      security: []
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: ok
  /register:
    post:
      summary: Register new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsRequest"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /login:
    post:
      summary: Login user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsRequest"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items:
    get:
      summary: Sync items since version
      security:
        - bearerAuth: []
      parameters:
        - name: since
          in: query
          description: Return items updated since this version (inclusive)
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        "200":
          description: Item list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemResponse"
        "400":
          description: Invalid since parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create or update an item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpsertRequest"
      responses:
        "200":
          description: Item upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict (version mismatch)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items/{id}:
    get:
      summary: Get item by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete item by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: version
          in: query
          required: false
          description: Expected item version (optimistic concurrency)
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        "200":
          description: Deleted item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          description: Invalid ID or version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict (version mismatch)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    CredentialsRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: alice
        password:
          type: string
          example: secret
    TokenResponse:
      type: object
      required: [token, expires_in_seconds]
      properties:
        token:
          type: string
        expires_in_seconds:
          type: integer
          format: int64
          example: 86400
    ItemUpsertRequest:
      type: object
      required: [type, payload]
      properties:
        id:
          type: string
          format: uuid
          description: Optional; if omitted a new ID is generated
        type:
          type: string
          description: Secret type
          enum: [login_password, text, binary, card]
        payload:
          type: string
          format: byte
          description: Base64-encoded encrypted payload
        version:
          type: integer
          format: int64
          description: Optional optimistic concurrency version
    ItemResponse:
      type: object
      required: [id, type, payload, deleted, version, updated_at]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [login_password, text, binary, card]
        payload:
          type: string
          format: byte
          description: Base64-encoded encrypted payload
        deleted:
          type: boolean
        version:
          type: integer
          format: int64
        updated_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string